added checking of exit codes

+++ tests.py
@@ -1,4 +1,6 @@
 import os
+import subprocess
+import sys
 tests = os.listdir("./tests/")
 print("running tests...")
 
@@ -7,7 +9,9 @@ os.system("chmod +x compile.sh")
 for test in tests:
     print(test)
     os.system("./compile.sh ./tests/" + test)
-    os.system("mono executable.exe")
+    executable = subprocess.call(["mono", "executable.exe"])
+    if not executable == 0:
+        sys.exit(executable)
     os.system("rm executable.exe")
 
 print("done")
